Minor fixes to type and name of getRenderResMultiplier

+++ src/ammonite/settings.cpp
@@ -153,7 +153,7 @@ namespace ammonite {
         graphics.renderResMultiplier = renderResMultiplier;
       }
 
-      int getRenderResMultiplier() {
+      float getRenderResMultiplier() {
         return graphics.renderResMultiplier;
       }
